smaller changes made
