+++ website/src/pages/edited-collections/edit-toc.page.server.ts
@@ -1,0 +1,24 @@
+import { serverSideClients } from "src/graphql"
+import * as Dailp from "src/graphql/dailp"
+
+export async function prerender() {
+  const { data, error } = await serverSideClients.dailp
+    .query<Dailp.EditedCollectionsQuery, Dailp.EditedCollectionsQueryVariables>(
+      Dailp.EditedCollectionsDocument,
+      {}
+    )
+    .toPromise()
+
+  if (!data || error) {
+    throw error
+  }
+
+  return data.allEditedCollections.flatMap((collection) => {
+    return {
+      url: `/collections/${collection.slug}/edit-toc`,
+      pageContext: {
+        pageProps: { collectionSlug: collection.slug },
+      },
+    }
+  })
+}
